Introduce repository dispatch event in update workflow
This commit introduces a repository dispatch event to trigger the cruft-update GitHub workflow manually. Also added a "Check for conflict markers" step in update checker workflow for different project templates including Rust, Terraform and Python. This step checks for merge conflict markers and prompts for manual resolution when conflict markers are detected. Furthermore, 'auto-merge' labels are now added to the automatically created PRs when new updates are detected with Cruft.